+++ ddtrace/contrib/flask/__init__.py
@@ -87,9 +87,6 @@ Example::
     # Override service name
     config.flask['service_name'] = 'custom-service-name'
 
-    # Report 401, and 403 responses as errors
-    config.http_server.error_statuses = '401,403'
-
 .. __: http://flask.pocoo.org/
 
 :ref:`All HTTP tags <http-tagging>` are supported for this integration.
