+++ ddtrace/contrib/internal/httplib/patch.py
@@ -91,7 +91,7 @@ def _wrap_request(func, instance, args, kwargs):
     if should_skip_request(pin, instance):
         return func_to_call(*args, **kwargs)
 
-    cfg = config.get_from(instance)
+    cfg = config._get_from(instance)
 
     try:
         # Create a new span and attach to this instance (so we can retrieve/update/close later on the response)
