+++ ddtrace/contrib/internal/requests/connection.py
@@ -75,7 +75,7 @@ def _wrap_send(func, instance, args, kwargs):
     hostname, path = _extract_hostname_and_path(url)
     host_without_port = hostname.split(":")[0] if hostname is not None else None
 
-    cfg = config.get_from(instance)
+    cfg = config._get_from(instance)
     service = None
     if cfg["split_by_domain"] and hostname:
         service = hostname
@@ -97,7 +97,7 @@ def _wrap_send(func, instance, args, kwargs):
 
         # Configure trace search sample rate
         # DEV: analytics enabled on per-session basis
-        cfg = config.get_from(instance)
+        cfg = config._get_from(instance)
         analytics_enabled = cfg.get("analytics_enabled")
         if analytics_enabled:
             span.set_tag(_ANALYTICS_SAMPLE_RATE_KEY, cfg.get("analytics_sample_rate", True))
