+++ ddtrace/contrib/internal/trace_utils.py
@@ -485,7 +485,7 @@ def set_http_meta(
             log.debug("failed to convert http status code %r to int", status_code)
         else:
             span.set_tag_str(http.STATUS_CODE, str(status_code))
-            if config.http_server.is_error_code(int_status_code):
+            if config._http_server.is_error_code(int_status_code):
                 span.error = 1
 
     if status_msg is not None:
