+++ ddtrace/settings/__init__.py
@@ -1,12 +1,12 @@
 from .._hooks import Hooks
-from .config import Config
+from ._config import Config
 from .exceptions import ConfigException
 from .http import HttpConfig
 from .integration import IntegrationConfig
 
 
 # Default global config
-_config = Config()
+_global_config = Config()
 
 __all__ = [
     "Config",
