+++ docs/advanced_usage.rst
@@ -396,7 +396,7 @@ Examples::
     from ddtrace import config
 
     # Global config
-    config.http.trace_query_string = True
+    # Set DD_TRACE_HTTP_CLIENT_TAG_QUERY_STRING environment variable to true/false
 
     # Integration level config, e.g. 'falcon'
     config.falcon.http.trace_query_string = True
@@ -471,14 +471,9 @@ structure like in the following example::
 Custom Error Codes
 ^^^^^^^^^^^^^^^^^^
 It is possible to have a custom mapping of which HTTP status codes are considered errors.
-By default, 500-599 status codes are considered errors.
-Configuration is provided both at the global level.
+By default, 500-599 status codes are considered errors. The default value can be overridden 
+by setting the ``DD_TRACE_HTTP_SERVER_ERROR_STATUSES`` environment variable.
 
-Examples::
-
-    from ddtrace import config
-
-    config.http_server.error_statuses = '500-599'
 
 Certain status codes can be excluded by providing a list of ranges. Valid options:
     - ``400-400``
