+++ tests/contrib/httplib/test_httplib_distributed.py
@@ -71,14 +71,14 @@ class TestHTTPLibDistributed(HTTPLibBaseMixin, TracerTestCase):
 
     def test_propagation_connection_true(self):
         conn = self.get_http_connection(SOCKET)
-        cfg = config.get_from(conn)
+        cfg = config._get_from(conn)
         cfg["distributed_tracing"] = True
         self.request(conn=conn)
         self.check_enabled()
 
     def test_propagation_connection_false(self):
         conn = self.get_http_connection(SOCKET)
-        cfg = config.get_from(conn)
+        cfg = config._get_from(conn)
         cfg["distributed_tracing"] = False
         self.request(conn=conn)
         self.check_disabled()
