+++ tests/contrib/requests/test_requests.py
@@ -281,7 +281,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_user_set_service_name(self):
         # ensure a service name set by the user has precedence
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["service_name"] = "clients"
         out = self.session.get(URL_200)
         assert out.status_code == 200
@@ -292,7 +292,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_user_set_service(self):
         # ensure a service name set by the user has precedence
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["service"] = "clients"
         out = self.session.get(URL_200)
         assert out.status_code == 200
@@ -333,7 +333,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
     def test_user_service_name_precedence(self):
         # ensure the user service name takes precedence over
         # the parent Span
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["service_name"] = "clients"
         with self.tracer.trace("parent.span", service="web"):
             out = self.session.get(URL_200)
@@ -349,7 +349,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
     def test_split_by_domain_with_ampersat(self):
         # Regression test for: https://github.com/DataDog/dd-trace-py/issues/4062
         # ensure a service name is generated by the domain name
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         # domain name should take precedence over monkey_service
         cfg["service_name"] = "monkey_service"
@@ -368,7 +368,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
     def test_split_by_domain(self):
         # ensure a service name is generated by the domain name
         # of the ongoing call
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         out = self.session.get(URL_200)
         assert out.status_code == 200
@@ -383,7 +383,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_split_by_domain_precedence(self):
         # ensure the split by domain has precedence all the time
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         cfg["service_name"] = "intake"
         out = self.session.get(URL_200)
@@ -400,7 +400,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
     def test_split_by_domain_wrong(self):
         # ensure the split by domain doesn't crash in case of a wrong URL;
         # in that case, no spans are created
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         with pytest.raises((MissingSchema, InvalidURL)):
             self.session.get("http:/some>thing")
@@ -411,7 +411,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_split_by_domain_remove_auth_in_url(self):
         # ensure that auth details are stripped from URL
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         out = self.session.get(f"http://user:pass@{HOST_AND_PORT}")
         assert out.status_code == 200
@@ -425,7 +425,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_split_by_domain_includes_port(self):
         # ensure that port is included if present in URL
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         out = self.session.get(f"http://{HOST_AND_PORT}")
         assert out.status_code == 200
@@ -439,7 +439,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
 
     def test_split_by_domain_includes_port_path(self):
         # ensure that port is included if present in URL but not path
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["split_by_domain"] = True
         out = self.session.get(f"http://{HOST_AND_PORT}/anything/v1/foo")
         assert out.status_code == 200
@@ -458,7 +458,7 @@ class TestRequests(BaseRequestTestCase, TracerTestCase):
         spans = self.pop_spans()
         assert spans[0].service == "override"
 
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["service"] = "override2"
         out = self.session.get(URL_200)
         assert out.status_code == 200
