+++ tests/contrib/requests/test_requests_distributed.py
@@ -77,7 +77,7 @@ class TestRequestsDistributed(BaseRequestTestCase, TracerTestCase):
 
     def test_propagation_true(self):
         # ensure distributed tracing can be enabled manually
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["distributed_tracing"] = True
         adapter = Adapter()
         self.session.mount("mock", adapter)
@@ -101,7 +101,7 @@ class TestRequestsDistributed(BaseRequestTestCase, TracerTestCase):
 
     def test_propagation_false(self):
         # ensure distributed tracing can be disabled manually
-        cfg = config.get_from(self.session)
+        cfg = config._get_from(self.session)
         cfg["distributed_tracing"] = False
         adapter = Adapter()
         self.session.mount("mock", adapter)
@@ -121,7 +121,7 @@ class TestRequestsDistributed(BaseRequestTestCase, TracerTestCase):
         with self.override_global_config(dict(_appsec_standalone_enabled=True, _asm_enabled=True)):
             assert asm_config._apm_opt_out
             self.tracer.enabled = False
-            cfg = config.get_from(self.session)
+            cfg = config._get_from(self.session)
             cfg["distributed_tracing"] = True
             adapter = Adapter()
             self.session.mount("mock", adapter)
@@ -149,7 +149,7 @@ class TestRequestsDistributed(BaseRequestTestCase, TracerTestCase):
             assert not asm_config._apm_opt_out
             self.tracer.enabled = False
 
-            cfg = config.get_from(self.session)
+            cfg = config._get_from(self.session)
             cfg["distributed_tracing"] = True
             adapter = Adapter()
             self.session.mount("mock", adapter)
